refactor(experimental): rename luVS module
