+++ docs/table-of-contents.json
@@ -218,7 +218,7 @@
           "api-reference/experimental/ludf",
           "api-reference/experimental/luxfilter",
           "api-reference/experimental/lutrace",
-          "api-reference/experimental/luvs",
+          "api-reference/experimental/graph-vector-similarity",
           "api-reference/experimental/g-buffer",
           "api-reference/experimental/deferred-lighting",
           "api-reference/experimental/clustered-lighting",
@@ -362,7 +362,7 @@
           "api-reference/experimental/ludf",
           "api-reference/experimental/luxfilter",
           "api-reference/experimental/lutrace",
-          "api-reference/experimental/luvs",
+          "api-reference/experimental/graph-vector-similarity",
           "api-reference/experimental/g-buffer",
           "api-reference/experimental/deferred-lighting",
           "api-reference/experimental/clustered-lighting",
