+++ modules/experimental/package.json
@@ -72,10 +72,10 @@
       "require": "./dist/luproj/index.cjs",
       "types": "./dist/luproj/index.d.ts"
     },
-    "./luvs": {
-      "import": "./dist/luvs/index.js",
-      "require": "./dist/luvs/index.cjs",
-      "types": "./dist/luvs/index.d.ts"
+    "./graph-vector-similarity": {
+      "import": "./dist/graph-vector-similarity/index.js",
+      "require": "./dist/graph-vector-similarity/index.cjs",
+      "types": "./dist/graph-vector-similarity/index.d.ts"
     },
     "./luproj/benchmarks": {
       "import": "./dist/luproj/benchmarks.js",
