+++ modules/experimental/src/graph-vector-similarity/embedding-matrix.ts
