+++ modules/experimental/test/index.ts
@@ -50,4 +50,4 @@ import './luraster';
 import './luxfilter';
 import './luproj/luproj.spec';
 import './luproj/projection-benchmark.spec';
-import './luvs/gpu-similarity-search.spec';
+import './graph-vector-similarity/gpu-similarity-search.spec';
